Refactor LSP server example with improved code structure
