+++ examples/BUILD.bazel
@@ -53,7 +53,6 @@ cc_binary(
     name = "pipe_lsp_server",
     srcs = [
         "lsp_example/server/lsp_server.cpp",
-        "lsp_example/utils.hpp",
     ],
     deps = ["//src:jsonrpc_lib"],
 )
